misc: move isobmff.js to parsers/containers
